requirements.txt
# Temel veri analizi kütüphaneleri
pandas>=1.5.0
numpy>=1.21.0
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# Görüntü işleme kütüphaneleri
opencv-python>=4.8.0
Pillow>=9.0.0
albumentations>=1.3.0

# Machine Learning kütüphaneleri
scikit-learn>=1.1.0

# Veri formatları ve dosya işleme
PyYAML>=6.0
lxml>=4.9.0

# Jupyter notebook desteği
jupyter>=1.0.0
ipykernel>=6.0.0

# Görselleştirme ve raporlama
reportlab>=3.6.0
jinja2>=3.0.0

# Progress bar ve utilities
tqdm>=4.64.0
pathlib

# Test kütüphaneleri
pytest>=7.0.0
pytest-cov>=4.0.0
